+++ src/FileStorage/Infrastructure/FileStorage.Infrastructure.DataAccess/Contexts/Images/Repositories/ImageRepository.cs
@@ -4,6 +4,7 @@ using FileStorage.Infrastructure.ObjectStorage;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using SixLabors.ImageSharp.PixelFormats;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -28,8 +29,8 @@ namespace FileStorage.Infrastructure.DataAccess.Contexts.Images.Repositories
         /// <inheritdoc />
         public async Task<ImageData> DownloadAsync(Guid id, CancellationToken cancellation)
         {
-            _logger.LogInformation("{0} -> Скачивание файла с ID: {1}",
-                nameof(DownloadAsync), id);
+            _logger.LogInformation("{0}:{1} -> Скачивание файла с ID: {2}",
+                nameof(ImageRepository), nameof(DownloadAsync), id);
 
             var fileBytes = await _objectStorage.GetData(id.ToString(), "images", cancellation);
             var file = new ImageData { Name = "name", Content = fileBytes, ContentType = "image/jpeg" };
@@ -40,16 +41,15 @@ namespace FileStorage.Infrastructure.DataAccess.Contexts.Images.Repositories
         /// <inheritdoc />
         public async Task<ImageShortInfo> GetInfoAsync(Guid id, CancellationToken cancellation)
         {
-            _logger.LogInformation("{0} -> Получение информации о файле с ID: {1}",
-                nameof(GetInfoAsync), id);
+            _logger.LogInformation("{0}:{1} -> Получение информации о файле с ID: {2}",
+                nameof(ImageRepository), nameof(GetInfoAsync), id);
 
             var objectStat = await _objectStorage.GetInfo(id.ToString(), "images", cancellation);
             if (objectStat == null)
             {
-                _logger.LogWarning($"Не найдено изображение с ID: {id}");
                 return null;
             }
-
+            
             var fileInfo = new ImageShortInfo { 
                 Name = objectStat.ObjectName,
                 CreatedAt = objectStat.LastModified, 
@@ -62,10 +62,25 @@ namespace FileStorage.Infrastructure.DataAccess.Contexts.Images.Repositories
         }
 
         /// <inheritdoc />
+        public async Task<bool> IsExists(Guid id, CancellationToken cancellation)
+        {
+            _logger.LogInformation("{0}:{1} -> Получение информации о наличии файла с ID: {2}",
+                nameof(ImageRepository), nameof(GetInfoAsync), id);
+
+            var objectStat = await _objectStorage.GetInfo(id.ToString(), "images", cancellation);
+            if (objectStat == null)
+            {
+                return false;
+            }
+
+            return true;
+        }
+
+        /// <inheritdoc />
         public async Task<Guid> UploadAsync(string contentType, byte[] bytes, CancellationToken cancellation)
         {
-            _logger.LogInformation("{0} -> Загрузка файла с содержимым: {1}",
-                nameof(UploadAsync), bytes);
+            _logger.LogInformation("{0}:{1} -> Загрузка файла с содержимым: {2}",
+                nameof(ImageRepository), nameof(UploadAsync), bytes);
 
             var fileName = Guid.NewGuid();
             var fileFolder = "images";
@@ -78,6 +93,15 @@ namespace FileStorage.Infrastructure.DataAccess.Contexts.Images.Repositories
         /// <inheritdoc />
         public Task DeleteAsync(Guid id, CancellationToken cancellation)
         {
+            _logger.LogInformation("{0}:{1} -> Удаление файла с ID: {2}",
+                nameof(ImageRepository), nameof(DeleteAsync), id);
+
+            var fileInfo = GetInfoAsync(id, cancellation);
+            if (fileInfo == null)
+            {
+                throw new KeyNotFoundException($"Не найден файл с ID: {id}");
+            }
+
             return _objectStorage.Delete(id.ToString(), "images", cancellation);
         }
     }
